+++ src/views/Faq.jsx
@@ -56,7 +56,6 @@ function Faq() {
       <section className="max-w-screen-xl mx-auto px-5 py-jj_60 text-deepDarkGray">
         <PageMainTitle pageTitleText="자주하는 질문" pageSubTitleText="궁금하신 내용을 검색해 주세요."></PageMainTitle>
         <SelectSearchFilter Collection="faq" handleReload={handleClickRefetch} option={searchOption} onChangeOption={setSearchOption} text={searchText} onChangeText={setSearchText}></SelectSearchFilter>
-        {/* <TableList collection="faq" field="faq" data={data} /> */}
         <TableListPagination searchText={searchText} searchOption={searchOption} error={error} collection="faq" field="faq" {...rest}></TableListPagination>
       </section>
     </>
